feat: add libinput scroll_button
This MR adds a option to set a button for when scroll_on_button_down is enabled as the scroll method. The user can define a different button to the enable the scroll event. By default the middle mouse button (274) is used, but with the `scroll_button` setting any other mouse button can be set, like the right mouse button (273) or a lateral button like back and forward buttons.